refactor(vue): replace createSpace function from Home.vue to LeftSidebar.vue
replaced createSpace and corresponding dependencies - createSpace, deleteBlankSpacename, isSpaceManagers, isSpaceUsers into LeftSidebar component
